refactor to float

+++ recipes-enhancer/src/main/java/me/auri/nutrients/NutriPicEnhancer.java
@@ -27,27 +27,27 @@ public class NutriPicEnhancer {
         // pizca
         nutrientDatabase.put("spaghetti",
                 new NutritionalFacts("spaghetti", 742,
-                        List.of(new Nutrient("Grasas", 3, "g"), new Nutrient("Hidratos de Carbono", 149.4, "g"),
+                        List.of(new Nutrient("Grasas", 3, "g"), new Nutrient("Hidratos de Carbono", 149.4f, "g"),
                                 new Nutrient("Proteinas", 26, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN)));
 
         nutrientDatabase.put("huevo",
                 new NutritionalFacts("huevo", 123,
-                        List.of(new Nutrient("Proteinas", 10.8, "g"), new Nutrient("Hidratos de Carbono", 0.6, "g"),
-                                new Nutrient("Grasas", 8.2, "g")),
+                        List.of(new Nutrient("Proteinas", 10.8f, "g"), new Nutrient("Hidratos de Carbono", 0.6f, "g"),
+                                new Nutrient("Grasas", 8.2f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGETARIAN, NutriPicResponse.HealthLabel.NO_SUGAR_ADDED,
                                 NutriPicResponse.HealthLabel.GLUTEN_FREE, NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
         nutrientDatabase.put("panceta",
-                new NutritionalFacts(
-                        "panceta", 393, List.of(new Nutrient("Proteinas", 13.7, "g"),
-                                new Nutrient("Hidratos de Carbono", 0, "g"), new Nutrient("Grasas", 37.1, "g")),
+                new NutritionalFacts("panceta", 393,
+                        List.of(new Nutrient("Proteinas", 13.7f, "g"), new Nutrient("Hidratos de Carbono", 0, "g"),
+                                new Nutrient("Grasas", 37.1f, "g")),
                         List.of(NutriPicResponse.HealthLabel.GLUTEN_FREE)));
 
         nutrientDatabase.put("queso parmesano",
-                new NutritionalFacts(
-                        "queso parmesano", 196, List.of(new Nutrient("Proteinas", 17.9, "g"),
-                                new Nutrient("Hidratos de Carbono", 1.6, "g"), new Nutrient("Grasas", 12.5, "g")),
+                new NutritionalFacts("queso parmesano", 196,
+                        List.of(new Nutrient("Proteinas", 17.9f, "g"), new Nutrient("Hidratos de Carbono", 1.6f, "g"),
+                                new Nutrient("Grasas", 12.5f, "g")),
                         List.of(NutriPicResponse.HealthLabel.GLUTEN_FREE)));
 
         nutrientDatabase.put("pimienta negra",
@@ -59,7 +59,7 @@ public class NutriPicEnhancer {
                                 NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
         nutrientDatabase.put("sal",
-                new NutritionalFacts("sal", 0, List.of(new Nutrient("Sodium", 147.5, "mg")),
+                new NutritionalFacts("sal", 0, List.of(new Nutrient("Sodium", 147.5f, "mg")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.GLUTEN_FREE,
                                 NutriPicResponse.HealthLabel.DAIRY_FREE)));
@@ -67,8 +67,8 @@ public class NutriPicEnhancer {
         // Pizza Margarita -> Harina:300g, Agua:200ml, Levadura:5g, Tomate:150g, Mozzarella:100g, Albahaca:5 hojas
         nutrientDatabase.put("harina",
                 new NutritionalFacts("harina", 1092,
-                        List.of(new Nutrient("Proteinas", 30.9, "g"), new Nutrient("Hidratos de Carbono", 228.9, "g"),
-                                new Nutrient("Grasas", 2.9, "g")),
+                        List.of(new Nutrient("Proteinas", 30.9f, "g"), new Nutrient("Hidratos de Carbono", 228.9f, "g"),
+                                new Nutrient("Grasas", 2.9f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN)));
 
         nutrientDatabase.put("agua",
@@ -81,52 +81,52 @@ public class NutriPicEnhancer {
 
         nutrientDatabase.put("levadura",
                 new NutritionalFacts("levadura", 16,
-                        List.of(new Nutrient("Proteinas", 2, "g"), new Nutrient("Hidratos de Carbono", 2.1, "g"),
-                                new Nutrient("Grasas", 0.4, "g")),
+                        List.of(new Nutrient("Proteinas", 2, "g"), new Nutrient("Hidratos de Carbono", 2.1f, "g"),
+                                new Nutrient("Grasas", 0.4f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
         nutrientDatabase.put("tomate",
                 new NutritionalFacts("tomate", 36,
-                        List.of(new Nutrient("Proteinas", 1.8, "g"), new Nutrient("Hidratos de Carbono", 8, "g"),
-                                new Nutrient("Grasas", 0.5, "g")),
+                        List.of(new Nutrient("Proteinas", 1.8f, "g"), new Nutrient("Hidratos de Carbono", 8, "g"),
+                                new Nutrient("Grasas", 0.5f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
         nutrientDatabase.put("mozarella",
                 new NutritionalFacts("mozarella", 299,
-                        List.of(new Nutrient("Proteinas", 22.2, "g"), new Nutrient("Hidratos de Carbono", 2.4, "g"),
-                                new Nutrient("Grasas", 22.1, "g")),
+                        List.of(new Nutrient("Proteinas", 22.2f, "g"), new Nutrient("Hidratos de Carbono", 2.4f, "g"),
+                                new Nutrient("Grasas", 22.1f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
         nutrientDatabase.put("albahaca",
                 new NutritionalFacts("albahaca", 2,
-                        List.of(new Nutrient("Proteinas", 0.3, "g"), new Nutrient("Hidratos de Carbono", 0.3, "g"),
-                                new Nutrient("Grasas", 0.1, "g")),
+                        List.of(new Nutrient("Proteinas", 0.3f, "g"), new Nutrient("Hidratos de Carbono", 0.3f, "g"),
+                                new Nutrient("Grasas", 0.1f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
         // Ensalada César -> Lechuga:1 unidad, Pollo:150g, Crutones:50g, Queso parmesano:30g, Aderezo César:50ml
         nutrientDatabase.put("lechuga",
                 new NutritionalFacts("lechuga", 26,
-                        List.of(new Nutrient("Proteinas", 2.7, "g"), new Nutrient("Hidratos de Carbono", 4.5, "g"),
-                                new Nutrient("Grasas", 0.4, "g")),
+                        List.of(new Nutrient("Proteinas", 2.7f, "g"), new Nutrient("Hidratos de Carbono", 4.5f, "g"),
+                                new Nutrient("Grasas", 0.4f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.DAIRY_FREE,
                                 NutriPicResponse.HealthLabel.GLUTEN_FREE)));
 
         nutrientDatabase.put("pollo",
                 new NutritionalFacts("pollo", 323,
-                        List.of(new Nutrient("Proteinas", 27.9, "g"), new Nutrient("Hidratos de Carbono", 0, "g"),
-                                new Nutrient("Grasas", 22.7, "g")),
+                        List.of(new Nutrient("Proteinas", 27.9f, "g"), new Nutrient("Hidratos de Carbono", 0, "g"),
+                                new Nutrient("Grasas", 22.7f, "g")),
                         List.of(NutriPicResponse.HealthLabel.NO_SUGAR_ADDED, NutriPicResponse.HealthLabel.DAIRY_FREE,
                                 NutriPicResponse.HealthLabel.GLUTEN_FREE)));
 
         nutrientDatabase.put("crutones",
                 new NutritionalFacts("crutones", 204,
-                        List.of(new Nutrient("Proteinas", 6, "g"), new Nutrient("Hidratos de Carbono", 36.8, "g"),
-                                new Nutrient("Grasas", 3.3, "g")),
+                        List.of(new Nutrient("Proteinas", 6, "g"), new Nutrient("Hidratos de Carbono", 36.8f, "g"),
+                                new Nutrient("Grasas", 3.3f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.DAIRY_FREE)));
 
@@ -141,22 +141,22 @@ public class NutriPicEnhancer {
 
         nutrientDatabase.put("calabaza",
                 new NutritionalFacts("calabaza", 130,
-                        List.of(new Nutrient("Proteinas", 5, "g"), new Nutrient("Hidratos de Carbono", 32.5, "g"),
-                                new Nutrient("Grasas", 0.5, "g")),
+                        List.of(new Nutrient("Proteinas", 5, "g"), new Nutrient("Hidratos de Carbono", 32.5f, "g"),
+                                new Nutrient("Grasas", 0.5f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.DAIRY_FREE, NutriPicResponse.HealthLabel.GLUTEN_FREE,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED)));
         nutrientDatabase.put("cebolla",
                 new NutritionalFacts("cebolla", 80,
-                        List.of(new Nutrient("Proteinas", 2.2, "g"), new Nutrient("Hidratos de Carbono", 18.7, "g"),
-                                new Nutrient("Grasas", 0.2, "g")),
+                        List.of(new Nutrient("Proteinas", 2.2f, "g"), new Nutrient("Hidratos de Carbono", 18.7f, "g"),
+                                new Nutrient("Grasas", 0.2f, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.DAIRY_FREE, NutriPicResponse.HealthLabel.GLUTEN_FREE,
                                 NutriPicResponse.HealthLabel.NO_SUGAR_ADDED)));
 
         nutrientDatabase.put("caldo de verduras",
                 new NutritionalFacts("caldo de verduras", 86,
-                        List.of(new Nutrient("Proteinas", 1.4, "g"), new Nutrient("Hidratos de Carbono", 1.2, "g"),
+                        List.of(new Nutrient("Proteinas", 1.4f, "g"), new Nutrient("Hidratos de Carbono", 1.2f, "g"),
                                 new Nutrient("Grasas", 1, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.DAIRY_FREE, NutriPicResponse.HealthLabel.GLUTEN_FREE,
@@ -164,7 +164,7 @@ public class NutriPicEnhancer {
 
         nutrientDatabase.put("jenjibre",
                 new NutritionalFacts("jenjibre", 4,
-                        List.of(new Nutrient("Proteinas", 0.1, "g"), new Nutrient("Hidratos de Carbono", 0.9, "g"),
+                        List.of(new Nutrient("Proteinas", 0.1f, "g"), new Nutrient("Hidratos de Carbono", 0.9f, "g"),
                                 new Nutrient("Grasas", 0, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGAN, NutriPicResponse.HealthLabel.VEGETARIAN,
                                 NutriPicResponse.HealthLabel.DAIRY_FREE, NutriPicResponse.HealthLabel.GLUTEN_FREE,
@@ -172,7 +172,7 @@ public class NutriPicEnhancer {
 
         nutrientDatabase.put("nata",
                 new NutritionalFacts("nata", 343,
-                        List.of(new Nutrient("Proteinas", 36.4, "g"), new Nutrient("Hidratos de Carbono", 2.9, "g"),
+                        List.of(new Nutrient("Proteinas", 36.4f, "g"), new Nutrient("Hidratos de Carbono", 2.9f, "g"),
                                 new Nutrient("Grasas", 0, "g")),
                         List.of(NutriPicResponse.HealthLabel.VEGETARIAN, NutriPicResponse.HealthLabel.GLUTEN_FREE)));
     }
@@ -192,7 +192,7 @@ public class NutriPicEnhancer {
                     Nutrient existing = nutrientMap.get(nutrient.getLabel());
                     if (existing != null) {
                         // If nutrient is already present in map, we add the quantities
-                        double newQuantity = existing.getQuantity() + nutrient.getQuantity();
+                        float newQuantity = existing.getQuantity() + nutrient.getQuantity();
                         nutrientMap.put(nutrient.getLabel(),
                                 new Nutrient(nutrient.getLabel(), newQuantity, nutrient.getUnit()));
                     } else {
